@@ -2,6 +2,9 @@
   "name": "nextjs-docs-ja",
   "version": "1.0.0",
   "private": true,
+  "engines": {
+    "node": ">=20.17.0"
+  },
   "scripts": {
     "docusaurus": "docusaurus",
     "start": "docusaurus start",
@@ -15,33 +18,31 @@
     "typecheck": "tsc",
     "lint": "npm run lint:docs",
     "lint:docs": "textlint --fix docs/**",
-    "portable": "npm-run-all --sequential portable:*",
-    "portable:clean": "rimraf dist",
-    "portable:pkg": "pkg -t node16-win-x64 -o dist/nextjs-docs-portable node_modules/http-server/bin/http-server",
-    "portable:build": "npm run build",
-    "portable:cp": "cpx \"build/**/*\" dist/public"
+    "kj:diff": "tsx ./scripts/konjacbot/diff.mts -o .kj-diff.txt",
+    "kj:ja": "tsx ./scripts/konjacbot/translate.mts -l ja .kj-diff.txt",
+    "kj:mod": "tsx ./scripts/konjacbot/mod.mts .kj-diff.txt",
+    "kj:img": "tsx ./scripts/konjacbot/dlimg.mts .kj-diff.txt",
+    "kj:pr": "tsx ./scripts/konjacbot/pr.mts"
   },
   "dependencies": {
-    "@docusaurus/core": "3.2.0",
-    "@docusaurus/module-type-aliases": "3.2.0",
-    "@docusaurus/preset-classic": "3.2.0",
-    "@docusaurus/theme-classic": "3.2.0",
+    "@docusaurus/core": "^3.5.2",
+    "@docusaurus/module-type-aliases": "^3.5.2",
+    "@docusaurus/preset-classic": "^3.5.2",
+    "@docusaurus/theme-classic": "^3.5.2",
     "@easyops-cn/docusaurus-search-local": "^0.40.1",
     "@mdx-js/react": "^3.0.1",
     "clsx": "^2.0.0",
     "docusaurus-plugin-image-zoom": "^2.0.0",
-    "prism-react-renderer": "^1.3.5",
+    "prism-react-renderer": "^2.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@tsconfig/docusaurus": "^2.0.3",
-    "cpx": "^1.5.0",
     "http-server": "^14.1.1",
-    "npm-run-all": "^4.1.5",
-    "pkg": "^5.8.1",
+    "openai": "^4.62.0",
+    "p-limit": "^6.1.0",
     "prettier": "3.2.5",
-    "rimraf": "^5.0.5",
     "textlint": "^14.0.4",
     "textlint-filter-rule-comments": "^1.2.2",
     "textlint-rule-preset-bengo4": "^2.2.2",
@@ -49,7 +50,9 @@
     "textlint-rule-preset-ja-technical-writing": "^10.0.1",
     "textlint-rule-prh": "^6.0.0",
     "textlint-rule-spellcheck-tech-word": "^5.0.0",
-    "typescript": "^5.4.3"
+    "tsx": "^4.19.1",
+    "typescript": "^5.6.2",
+    "zx": "^8.1.7"
   },
   "browserslist": {
     "production": [
